// prisma/schemas/inputTypeSchemas/GroupHeadUncheckedCreateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { GroupHeadLevelSchema } from './GroupHeadLevelSchema';

export const GroupHeadUncheckedCreateInputSchema: z.ZodType<Prisma.GroupHeadUncheckedCreateInput> = z.object({
  id: z.string(),
  tgID: z.bigint(),
  group: z.number().int(),
  level: z.lazy(() => GroupHeadLevelSchema),
  createdAt: z.coerce.date().optional(),
  updatedAt: z.coerce.date().optional()
}).strict();

export default GroupHeadUncheckedCreateInputSchema;